+++ src/main/java/org/terracotta/ehcache/testing/sequencegenerator/RandomSequenceGenerator.java
@@ -5,11 +5,11 @@ import java.util.Random;
 public class RandomSequenceGenerator implements SequenceGenerator {
 
   private final Distribution distribution;
-  private final int minimum;
-  private final int maximum;
-  private final int width;
-  
-  public RandomSequenceGenerator(Distribution distribution, int min, int max, int width) {
+  private final long minimum;
+  private final long maximum;
+  private final long width;
+
+  public RandomSequenceGenerator(Distribution distribution, long min, long max, long width) {
     this.distribution = distribution;
     this.minimum = min;
     this.maximum = max;
@@ -19,23 +19,23 @@ public class RandomSequenceGenerator implements SequenceGenerator {
   public Sequence createSequence() {
     return new RandomSequence(distribution, minimum, maximum, width);
   }
-  
+
   static class RandomSequence implements Sequence {
 
     private final Random rndm = new Random();
     private final Distribution distribution;
-    private final int minimum;
-    private final int maximum;
-    private final int width;
-    
-    public RandomSequence(Distribution distribution, int minimum, int maximum, int width) {
+    private final long minimum;
+    private final long maximum;
+    private final long width;
+
+    public RandomSequence(Distribution distribution, long minimum, long maximum, long width) {
       this.distribution = distribution;
       this.minimum = minimum;
       this.maximum = maximum;
       this.width = width;
     }
-    
-    public int next() {
+
+    public long next() {
       return distribution.generate(rndm, minimum, maximum, width);
     }
   }
