+++ src/main/java/org/terracotta/ehcache/testing/sequencegenerator/SequenceGenerator.java
@@ -3,8 +3,8 @@ package org.terracotta.ehcache.testing.sequencegenerator;
 public interface SequenceGenerator {
 
   public Sequence createSequence();
-  
+
   public interface Sequence {
-    public int next();
+    public long next();
   }
 }
